F82 Redirect after login.
To allow for redirecting after logging in to the page
that user was trying to enter, a small store modifi-
cation has been introduced. Now one show be taken back
to the page that he or she was trying to display after
login. Both login and registration+login flows should
work.